🎨 refactor : Extract components
